requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
async-timeout==5.0.1
attrs==25.4.0
beautifulsoup4==4.14.3
cachetools==6.2.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudscraper==1.2.71
colorama==0.4.6
cryptography==46.0.3
deep-translator==1.11.4
deprecation==2.1.0
edge-tts==7.2.3
exceptiongroup==1.3.1
fastapi==0.124.4
feedparser==6.0.12
frozenlist==1.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
multidict==6.7.0
opentele==1.15.1
outcome==1.3.0.post0
packaging==25.0
playwright==1.57.0
postgrest==2.25.1
praw==7.8.1
prawcore==2.4.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
PyJWT==2.10.1
pyparsing==3.2.5
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.2
Pyrogram==2.0.106
PySocks==1.7.1
pyTelegramBotAPI==4.29.1
python-dotenv==1.2.1
python-socks==2.8.0
pytz==2025.2
realtime==2.25.1
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
schedule==1.2.2
sgmllib3k==1.0.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SpeechRecognition==3.14.4
starlette==0.50.0
storage3==2.25.1
StrEnum==0.4.15
supabase==2.25.1
supabase-auth==2.25.1
supabase-functions==2.25.1
tabulate==0.9.0
Telethon==1.42.0
TgCrypto==1.2.5
tqdm==4.67.1
trio==0.32.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
update-checker==0.18.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
websocket-client==1.9.0
websockets==15.0.1
wsproto==1.3.2
yarl==1.22.0
